Fixed a bug and made some major changed
Added field level validationas and error messages. Changed entity IDs from INTEGER to UUID

+++ src/main/server/helpers/FormattedLog.java
@@ -1,7 +1,7 @@
 package helpers;
 
 public record FormattedLog(
-        int logId,
+        long logId,
         String createdAtDate,
         String createdAtTime,
         String endPoint,
